jsf/el fix
git-svn-id: svn://svn.caucho.com/resin/trunk@7399 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/kernel/src/com/caucho/el/Expr.java
@@ -796,7 +796,7 @@ public abstract class Expr extends ValueExpression {
       return 0;
     else if (value instanceof String) {
       try {
-	return (long) Double.parseDouble((String) value);
+	return Long.parseLong((String) value);
       } catch (Exception e) {
 	throw new ELException(e);
       }
